+++ CHANGELOG.md
@@ -20,6 +20,7 @@ This changelog is effective from the 2025 releases.
 * Script `generate_example.sh` to generate documentation pages from notebook examples
 * GitHub workflows for CI and publishing to PyPI
 * Build using `pyproject.toml`, addition of extras groups to install optional dependencies
+* `LogManager` and `TextLogger` to manage log files and console logging
 
 ### Changed
 * Functions for optional packages (e.g. RDKit, ASE) are available even when these packages are not installed, but will raise an `MissingOptionalPackageError` when called
@@ -30,6 +31,7 @@ This changelog is effective from the 2025 releases.
 * Supercell and RDKit properties are no longer serialized to AMS input
 * Restructuring of examples and conversion of various examples to notebooks
 * Support for `networkx>=3` and `ase>=3.23`
+* Use standard library logger for `log` function
 
 ### Fixed
 * `Molecule.properties.charge` is a numeric instead of string type when loading molecule from a file